// src/types/states.ts
/**
 * @description
 * States enum
 * @enum {string}
 */
export enum STATES {
  ACTIVE = 'ACTIVE',
  COMPLETED = 'COMPLETED',
  DEFAULT = 'DEFAULT',
  DEFAULT_SELECTED = 'DEFAULT_SELECTED',
  DEFAULT_UNSELECTED = 'DEFAULT_UNSELECTED',
  DISABLED = 'DISABLED',
  DISABLED_SELECTED = 'DISABLED_SELECTED',
  DISABLED_UNSELECTED = 'DISABLED_UNSELECTED',
  ERROR = 'ERROR',
  FILLED = 'FILLED',
  FOCUS = 'FOCUS',
  HOVER = 'HOVER',
  PRESSED = 'PRESSED',
  SELECTED = 'SELECTED',
  LOADING = 'LOADING',
  SUCCESS = 'SUCCESS',
  WARNING = 'WARNING',
  VISITED = 'VISITED',
}